# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: NovaMind API
  version: 0.1.0
  description: |
    NovaMind V2 后端 API。所有响应统一为 `{ data, error, trace_id }`：
    - 成功：`data` 有值、`error` 为 null
    - 失败：`data` 为 null、`error` 含 `code` / `message`
    `trace_id` 与响应头 `X-Request-Id` 一致，可用于排查日志。
servers:
  - url: /api/v1
    description: 当前服务
...
paths:
...
  /model-providers:
    get:
      tags: [ai]
      summary: 模型配置列表
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ModelProviderList"
    post:
      tags: [ai]
      summary: 新增模型配置
      description: api_key 会用 NOVAMIND_SECRET 加密入库；接口永不返回密钥，只返回 has_api_key。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelProviderRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ModelProvider"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /model-providers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [ai]
      summary: 模型配置详情
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ModelProvider"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ai]
      summary: 更新模型配置
      description: api_key 留空表示保持原密钥不变。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelProviderRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ModelProvider"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [ai]
      summary: 删除模型配置
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /model-providers/{id}/default:
    post:
      tags: [ai]
      summary: 设为默认模型
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 设置成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ModelProvider"
        "404":
          $ref: "#/components/responses/NotFound"
  /model-providers/{id}/test:
    post:
      tags: [ai]
      summary: 测试模型配置
      description: 用一条最小请求真实调用上游（会消耗极少量 token）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 测试结果（ok=false 表示调用失败，error_message 给出原因）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ModelProviderTest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    BadRequest:
      description: 参数不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    Conflict:
      description: 资源冲突（例如该工程已存在原著）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
  schemas:
    Envelope:
      type: object
      required: [data, error, trace_id]
      properties:
        data:
          nullable: true
          description: 成功时的业务数据；失败时为 null
        error:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/APIError"
        trace_id:
          type: string
          description: 与响应头 X-Request-Id 一致
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: PROJECT_NOT_FOUND
        message:
          type: string
          example: 项目不存在
        details:
          type: object
          additionalProperties: true
          nullable: true
...
    ModelProviderRequest:
      type: object
      required: [name, provider, api_base, model_name]
      properties:
        name: { type: string, maxLength: 120, example: DeepSeek }
        provider: { type: string, enum: [OPENAI_COMPATIBLE, ANTHROPIC] }
        api_base: { type: string, example: https://api.deepseek.com/v1 }
        api_key: { type: string, description: 明文只在请求里出现，服务端加密存储 }
        model_name: { type: string, example: deepseek-chat }
        purpose: { type: string, enum: [chat, embedding, both], default: chat }
        temperature: { type: number, minimum: 0, maximum: 2, default: 0.7 }
        max_tokens: { type: integer, default: 4096 }
        timeout_sec: { type: integer, default: 120 }
        enabled: { type: boolean, default: true }
        is_default: { type: boolean, default: false }
        notes: { type: string }
    ModelProvider:
      allOf:
        - $ref: "#/components/schemas/ModelProviderRequest"
        - type: object
          required: [id, has_api_key, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            api_key: { type: string, readOnly: true, description: 永远不返回 }
            has_api_key: { type: boolean, description: 是否已保存密钥（不暴露任何密钥片段） }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    ModelProviderList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModelProvider"
        total: { type: integer }
    ModelProviderTest:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        model: { type: string }
        reply: { type: string }
        latency_ms: { type: integer, format: int64 }
        total_tokens: { type: integer }
        error_message: { type: string }
...